backend/requirements.txt
fastapi
uvicorn[standard]
geopandas
shapely>=2.0
python-dotenv
